Move `*Version()` functions to version.h/cpp
