+++ src/qt/utilitydialog.cpp
@@ -21,10 +21,10 @@
 #include "clientmodel.h"
 #include "guiutil.h"
 
-#include "clientversion.h"
 #include "init.h"
 #include "util.h"
 #include "net.h"
+#include "version.h"
 
 #include <QLabel>
 #include <QFont>
