+++ src/rpcprotocol.cpp
@@ -6,6 +6,7 @@
 #include "rpcprotocol.h"
 
 #include "util.h"
+#include "version.h"
 
 #include <stdint.h>
